Split entity classes by responsibility
