+++ src/sv/entities/__init__.py
@@ -1,2 +1,6 @@
 """Player and enemy entities."""
-from .entity import Player, Skeleton
+from .base import Entity
+from .enemy import Enemy, Skeleton
+from .player import Player
+
+__all__ = ["Enemy", "Entity", "Player", "Skeleton"]
